refactor: Alembic Migrationen zu Initial-Schema konsolidiert
14 Migrationen durch eine saubere Initial-Migration ersetzt.
Keine Produktions-DB betroffen.

Entfernte redundante Änderungen:
- FreezeTimeConfig (hinzugefügt + entfernt)
- ItemCategory junction table (hinzugefügt + entfernt)
- expiry_date auf item (hinzugefügt + entfernt)
- freeze_time_months auf category (hinzugefügt + entfernt)
- failed_login_attempts auf users (hinzugefügt + entfernt)

Finale Tabellen: users, category, category_shelf_life, location,
item, login_attempt, system_settings, withdrawal
