--- alembic/versions/05ee025b32ea_add_category_table.py
@@ -1,49 +1,0 @@
-"""add Category table
-
-Revision ID: 05ee025b32ea
-Revises: e427f74927c3
-Create Date: 2025-11-24 02:48:22.324015
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel
-from typing import Sequence
-from typing import Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "05ee025b32ea"
-down_revision: Union[str, Sequence[str], None] = "e427f74927c3"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "category",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("name", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column("color", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.Column("freeze_time_months", sa.Integer(), nullable=True),
-        sa.Column("created_at", sa.DateTime(), nullable=False),
-        sa.Column("created_by", sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["created_by"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_category_name"), "category", ["name"], unique=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_category_name"), table_name="category")
-    op.drop_table("category")
-    # ### end Alembic commands ###
