--- alembic/versions/4a9f912e1477_add_item_and_itemcategory_tables.py
@@ -1,77 +1,0 @@
-"""add Item and ItemCategory tables
-
-Revision ID: 4a9f912e1477
-Revises: cce4a4601eec
-Create Date: 2025-11-24 09:43:43.891026
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel
-from typing import Sequence
-from typing import Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "4a9f912e1477"
-down_revision: Union[str, Sequence[str], None] = "cce4a4601eec"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "item",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("product_name", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column("best_before_date", sa.Date(), nullable=False),
-        sa.Column("freeze_date", sa.Date(), nullable=True),
-        sa.Column("expiry_date", sa.Date(), nullable=False),
-        sa.Column("quantity", sa.Float(), nullable=False),
-        sa.Column("unit", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column(
-            "item_type", sa.Enum("TINNED", "JARRED", "FROZEN", "CHILLED", "AMBIENT", name="itemtype"), nullable=False
-        ),
-        sa.Column("location_id", sa.Integer(), nullable=False),
-        sa.Column("notes", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.Column("is_consumed", sa.Boolean(), nullable=False),
-        sa.Column("created_at", sa.DateTime(), nullable=False),
-        sa.Column("created_by", sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["created_by"],
-            ["users.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["location_id"],
-            ["location.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_item_product_name"), "item", ["product_name"], unique=False)
-    op.create_table(
-        "item_category",
-        sa.Column("item_id", sa.Integer(), nullable=False),
-        sa.Column("category_id", sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["category_id"],
-            ["category.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["item_id"],
-            ["item.id"],
-        ),
-        sa.PrimaryKeyConstraint("item_id", "category_id"),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("item_category")
-    op.drop_index(op.f("ix_item_product_name"), table_name="item")
-    op.drop_table("item")
-    # ### end Alembic commands ###
