--- alembic/versions/7133d257dd27_add_location_table.py
@@ -1,50 +1,0 @@
-"""add Location table
-
-Revision ID: 7133d257dd27
-Revises: 05ee025b32ea
-Create Date: 2025-11-24 09:25:21.628495
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel
-from typing import Sequence
-from typing import Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "7133d257dd27"
-down_revision: Union[str, Sequence[str], None] = "05ee025b32ea"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "location",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("name", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column("location_type", sa.Enum("FROZEN", "CHILLED", "AMBIENT", name="locationtype"), nullable=False),
-        sa.Column("description", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.Column("is_active", sa.Boolean(), nullable=False),
-        sa.Column("created_at", sa.DateTime(), nullable=False),
-        sa.Column("created_by", sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["created_by"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_location_name"), "location", ["name"], unique=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_location_name"), table_name="location")
-    op.drop_table("location")
-    # ### end Alembic commands ###
