--- alembic/versions/95be42737cf2_add_withdrawal_table.py
@@ -1,43 +1,0 @@
-"""add withdrawal table
-
-Revision ID: 95be42737cf2
-Revises: 83f2c87ee2a6
-Create Date: 2025-11-30 20:25:51.146956
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '95be42737cf2'
-down_revision: Union[str, Sequence[str], None] = '83f2c87ee2a6'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('withdrawal',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('item_id', sa.Integer(), nullable=False),
-    sa.Column('quantity', sa.Float(), nullable=False),
-    sa.Column('withdrawn_at', sa.DateTime(), nullable=False),
-    sa.Column('withdrawn_by', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['item_id'], ['item.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['withdrawn_by'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_withdrawal_item_id'), 'withdrawal', ['item_id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_withdrawal_item_id'), table_name='withdrawal')
-    op.drop_table('withdrawal')
-    # ### end Alembic commands ###
