--- alembic/versions/c800634c7cf4_add_user_preferences_and_system_settings.py
@@ -1,43 +1,0 @@
-"""add_user_preferences_and_system_settings
-
-Revision ID: c800634c7cf4
-Revises: 5c90fbaec136
-Create Date: 2025-11-26 23:28:31.294189
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import sqlite
-
-# revision identifiers, used by Alembic.
-revision: str = 'c800634c7cf4'
-down_revision: Union[str, Sequence[str], None] = '5c90fbaec136'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # Create system_settings table for admin-configurable defaults
-    op.create_table('system_settings',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('key', sa.String(), nullable=False),
-    sa.Column('value', sa.String(), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), nullable=False),
-    sa.Column('updated_by', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['updated_by'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_system_settings_key'), 'system_settings', ['key'], unique=True)
-
-    # Add preferences JSON column to users table for personal smart defaults
-    op.add_column('users', sa.Column('preferences', sqlite.JSON(), nullable=True))
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    op.drop_column('users', 'preferences')
-    op.drop_index(op.f('ix_system_settings_key'), table_name='system_settings')
-    op.drop_table('system_settings')
