--- alembic/versions/cce4a4601eec_add_freezetimeconfig_table.py
@@ -1,53 +1,0 @@
-"""add FreezeTimeConfig table
-
-Revision ID: cce4a4601eec
-Revises: 7133d257dd27
-Create Date: 2025-11-24 09:36:07.328283
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-from typing import Sequence
-from typing import Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "cce4a4601eec"
-down_revision: Union[str, Sequence[str], None] = "7133d257dd27"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "freeze_time_config",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("category_id", sa.Integer(), nullable=True),
-        sa.Column(
-            "item_type", sa.Enum("TINNED", "JARRED", "FROZEN", "CHILLED", "AMBIENT", name="itemtype"), nullable=False
-        ),
-        sa.Column("freeze_time_months", sa.Integer(), nullable=False),
-        sa.Column("created_at", sa.DateTime(), nullable=False),
-        sa.Column("created_by", sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(
-            ["category_id"],
-            ["category.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["created_by"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("category_id", "item_type", name="uq_category_item_type"),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("freeze_time_config")
-    # ### end Alembic commands ###
