+++ alembic/versions/d34a94a28640_initial_schema.py
@@ -1,0 +1,192 @@
+"""initial_schema
+
+Revision ID: d34a94a28640
+Revises:
+Create Date: 2025-11-30 22:22:51.420744
+
+"""
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import sqlite
+import sqlmodel
+from typing import Sequence
+from typing import Union
+
+
+# revision identifiers, used by Alembic.
+revision: str = "d34a94a28640"
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "login_attempt",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("ip_address", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+        sa.Column("fail_count", sa.Integer(), nullable=False),
+        sa.Column("last_attempt", sa.DateTime(), nullable=False),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_login_attempt_ip_address"), "login_attempt", ["ip_address"], unique=False)
+    op.create_table(
+        "users",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("username", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+        sa.Column("password_hash", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+        sa.Column("email", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+        sa.Column("role", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.Column("locked_until", sa.DateTime(), nullable=True),
+        sa.Column("remember_token", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+        sa.Column("preferences", sqlite.JSON(), nullable=True),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("last_login", sa.DateTime(), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_users_email"), "users", ["email"], unique=True)
+    op.create_index(op.f("ix_users_username"), "users", ["username"], unique=True)
+    op.create_table(
+        "category",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("name", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+        sa.Column("color", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+        sa.Column("sort_order", sa.Integer(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("created_by", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["created_by"],
+            ["users.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_category_name"), "category", ["name"], unique=True)
+    op.create_table(
+        "location",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("name", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+        sa.Column("location_type", sa.Enum("FROZEN", "CHILLED", "AMBIENT", name="locationtype"), nullable=False),
+        sa.Column("color", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+        sa.Column("description", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+        sa.Column("is_active", sa.Boolean(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("created_by", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["created_by"],
+            ["users.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_location_name"), "location", ["name"], unique=True)
+    op.create_table(
+        "system_settings",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("key", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+        sa.Column("value", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+        sa.Column("updated_at", sa.DateTime(), nullable=False),
+        sa.Column("updated_by", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["updated_by"],
+            ["users.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_system_settings_key"), "system_settings", ["key"], unique=True)
+    op.create_table(
+        "category_shelf_life",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("category_id", sa.Integer(), nullable=False),
+        sa.Column("storage_type", sa.Enum("FROZEN", "CHILLED", "AMBIENT", name="storagetype"), nullable=False),
+        sa.Column("months_min", sa.Integer(), nullable=False),
+        sa.Column("months_max", sa.Integer(), nullable=False),
+        sa.Column("source_url", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["category_id"],
+            ["category.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+        sa.UniqueConstraint("category_id", "storage_type", name="uq_category_storage"),
+    )
+    op.create_table(
+        "item",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("product_name", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+        sa.Column("best_before_date", sa.Date(), nullable=False),
+        sa.Column("freeze_date", sa.Date(), nullable=True),
+        sa.Column("quantity", sa.Float(), nullable=False),
+        sa.Column("unit", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+        sa.Column(
+            "item_type",
+            sa.Enum(
+                "PURCHASED_FRESH",
+                "PURCHASED_FROZEN",
+                "PURCHASED_THEN_FROZEN",
+                "HOMEMADE_FROZEN",
+                "HOMEMADE_PRESERVED",
+                name="itemtype",
+            ),
+            nullable=False,
+        ),
+        sa.Column("location_id", sa.Integer(), nullable=False),
+        sa.Column("category_id", sa.Integer(), nullable=True),
+        sa.Column("notes", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
+        sa.Column("is_consumed", sa.Boolean(), nullable=False),
+        sa.Column("created_at", sa.DateTime(), nullable=False),
+        sa.Column("created_by", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["category_id"],
+            ["category.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["created_by"],
+            ["users.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["location_id"],
+            ["location.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_item_product_name"), "item", ["product_name"], unique=False)
+    op.create_table(
+        "withdrawal",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("item_id", sa.Integer(), nullable=False),
+        sa.Column("quantity", sa.Float(), nullable=False),
+        sa.Column("withdrawn_at", sa.DateTime(), nullable=False),
+        sa.Column("withdrawn_by", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(["item_id"], ["item.id"], ondelete="CASCADE"),
+        sa.ForeignKeyConstraint(
+            ["withdrawn_by"],
+            ["users.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_withdrawal_item_id"), "withdrawal", ["item_id"], unique=False)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f("ix_withdrawal_item_id"), table_name="withdrawal")
+    op.drop_table("withdrawal")
+    op.drop_index(op.f("ix_item_product_name"), table_name="item")
+    op.drop_table("item")
+    op.drop_table("category_shelf_life")
+    op.drop_index(op.f("ix_system_settings_key"), table_name="system_settings")
+    op.drop_table("system_settings")
+    op.drop_index(op.f("ix_location_name"), table_name="location")
+    op.drop_table("location")
+    op.drop_index(op.f("ix_category_name"), table_name="category")
+    op.drop_table("category")
+    op.drop_index(op.f("ix_users_username"), table_name="users")
+    op.drop_index(op.f("ix_users_email"), table_name="users")
+    op.drop_table("users")
+    op.drop_index(op.f("ix_login_attempt_ip_address"), table_name="login_attempt")
+    op.drop_table("login_attempt")
+    # ### end Alembic commands ###
