--- alembic/versions/e427f74927c3_initial_migration_user_table.py
@@ -1,52 +1,0 @@
-"""Initial migration - User table
-
-Revision ID: e427f74927c3
-Revises:
-Create Date: 2025-11-24 02:34:27.126303
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-import sqlmodel
-from typing import Sequence
-from typing import Union
-
-
-# revision identifiers, used by Alembic.
-revision: str = "e427f74927c3"
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "users",
-        sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("username", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column("password_hash", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column("email", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column("role", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
-        sa.Column("is_active", sa.Boolean(), nullable=False),
-        sa.Column("failed_login_attempts", sa.Integer(), nullable=False),
-        sa.Column("locked_until", sa.DateTime(), nullable=True),
-        sa.Column("remember_token", sqlmodel.sql.sqltypes.AutoString(), nullable=True),
-        sa.Column("created_at", sa.DateTime(), nullable=False),
-        sa.Column("last_login", sa.DateTime(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_index(op.f("ix_users_email"), "users", ["email"], unique=True)
-    op.create_index(op.f("ix_users_username"), "users", ["username"], unique=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f("ix_users_username"), table_name="users")
-    op.drop_index(op.f("ix_users_email"), table_name="users")
-    op.drop_table("users")
-    # ### end Alembic commands ###
